requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
gunicorn>=21.2.0
Faker>=22.0.0
httpx>=0.25.0
flasgger>=0.9.7
pytest-cov>=5.0.0

